complete experiments

+++ FraudForaging/loop_functions/generate_source_pos.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python3
 import os, math
 import random
-os.remove('loop_functions/source_pos.txt')
+if os.path.exists('loop_functions/source_pos.txt'):
+    os.remove('loop_functions/source_pos.txt')
 from loop_function_params import *
-num_faulty = int(os.environ["NUM2"])
+num_malicious = int(os.environ["NUM2"])
 def distance(list1, list2):
     """Distance between two vectors."""
     squares = [(p-q) ** 2 for p, q in zip(list1, list2)]
@@ -29,7 +30,7 @@ def load(path):
     return fs_list
 
 fs_list=[]
-while len(fs_list)<(generic_params['num_food_source']+num_faulty):
+while len(fs_list)<(generic_params['num_food_source']+num_malicious):
     print(len(fs_list))
     fs=[0,0]
     minIntSrcDist = 0.3
